Move to osu.Framework project
